% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmejd723.R
\docType{data}
\name{altmejd723}
\alias{altmejd723}
\title{Replication database without the SSRP}
\format{
A data frame with 131 rows (study effects) and the 7 columns of
\link{altmejd}.
}
\source{
Altmejd A, Dreber A, Forsell E, Huber J, Imai T, et al. (2019)
Predicting the replicability of social science lab experiments. PLOS ONE
14(12): e0225826. \url{https://doi.org/10.1371/journal.pone.0225826}

The data and more information are in their OSF repository:
\url{https://osf.io/4fn73/}.
}
\usage{
altmejd723
}
\description{
A subset of the replication database of Altmejd et al. (2019), used only
to build \link{altmejd}. It holds four of the five projects: Many Labs 1 and 3,
the Replication Project Psychology and the Experimental Economics
Replication Project. The Social Science Replication Project is in
\link{ssrp723}.
}
\seealso{
Other datasets:
\code{\link{altmejd}},
\code{\link{altmejd_splits}},
\code{\link{hypergrid}},
\code{\link{hypernested}},
\code{\link{ssrp723}}
}
\concept{datasets}
\keyword{datasets}
