% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarimpAda.R
\name{plotVarimpAda}
\alias{plotVarimpAda}
\title{Plot AdaBoost Feature Importance}
\usage{
plotVarimpAda(imp_df, top_n = 15)
}
\arguments{
\item{imp_df}{The data frame output from \code{\link[adatutor]{varimpAda}}.}

\item{top_n}{The number of top variables to display (defaults to 15).}
}
\description{
Generates a horizontal bar chart of the top N most important features.
}
