% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrp723.R
\docType{data}
\name{ssrp723}
\alias{ssrp723}
\title{Replication database, SSRP only}
\format{
A data frame with 21 rows (study effects) and the 7 columns of
\link{altmejd}.
}
\source{
Altmejd A, Dreber A, Forsell E, Huber J, Imai T, et al. (2019)
Predicting the replicability of social science lab experiments. PLOS ONE
14(12): e0225826. \url{https://doi.org/10.1371/journal.pone.0225826}

The data and more information are in their OSF repository:
\url{https://osf.io/4fn73/}.
}
\usage{
ssrp723
}
\description{
A subset of the replication database of Altmejd et al. (2019), used only
to build \link{altmejd}. It holds one project: the Social Science Replication
Project (SSRP).
}
\seealso{
Other datasets:
\code{\link{altmejd}},
\code{\link{altmejd723}},
\code{\link{altmejd_splits}},
\code{\link{hypergrid}},
\code{\link{hypernested}}
}
\concept{datasets}
\keyword{datasets}
