% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAda.R
\name{testAda}
\alias{testAda}
\title{Generate Predictions or Retrodictions From a Set of (Adaptively) Boosted Trees and Model Weights}
\usage{
testAda(fit, data, input_checks = TRUE, verbose = TRUE)
}
\arguments{
\item{fit}{A model fitted with \code{\link[adatutor]{trainAda}}. This is a
list of boosted trees and theirs weights. More specifically, each element
of the list contains a weak learner (e.g., a decision stump) and its
corresponding weight.}

\item{data}{A data frame containing the test set for predictions. The
structure should match the training set used to generate \code{h}.}

\item{input_checks}{A logical value indicating whether to perform input
validation checks. Defaults to \code{TRUE}.}

\item{verbose}{A logical value specifying whether to display progress
messages and animations. Defaults to \code{TRUE}.}
}
\value{
A numeric vector containing the final predictions from the AdaBoost
model. The predictions are in the form of \code{-1} or \code{1},
corresponding to binary classification outcomes.
}
\description{
This function implements the testing phase of the AdaBoost
algorithm. It extracts the adaptively boosted weak learners (e.g.
classification stumps) and their corresponding weights to combine them into
the weighted sum \deqn{H(\mathbf{x}) = \sum_{i=1}^Ta_th_t(\mathbf x),}
which yields AdaBoost's predictions or retrodictions on a given set.
}
\details{
The function proceeds as follows:
\enumerate{
\item If \code{input_checks} is \code{TRUE}, basic input validation is
performed to ensure that \code{h} is a valid AdaBoost model and \code{data}
is appropriate for predictions.
\item Progress messages and animations are shown if \code{verbose} is
\code{TRUE}.
\item Boosted weak learners (\code{h}) and their weights (\code{a}) are extracted
from \code{fit}.
\item Predictions or retrodictions are generated for each weak learner
using the test set.
\item Individual predictions are combined using the weights from the
AdaBoost model to produce the final predictions.
}
}
\examples{
# Example usage:
# Assume `fit` is a trained model and `test` is a data frame.
# ypred <- testAda(fit, test)

}
