% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\alias{check_df}
\alias{check_eta}
\alias{check_length}
\alias{check_list}
\alias{check_numeric}
\alias{check_train}
\alias{check_prop}
\title{Miscellaneous Functions for Input Validation}
\usage{
check_df(x)

check_eta(x)

check_length(x)

check_list(x)

check_numeric(x)

check_train(trainnme, testnme)

check_prop(x)
}
\arguments{
\item{x}{A generic input to be validated. The expected type depends on the
specific function.}

\item{trainnme}{A character string representing the name of the training
set.}

\item{testnme}{A character string representing the name of the test
set.}
}
\value{
All functions are called for their side effects. If the input is
invalid, an error, or warning is generated with helpful suggestions for
troubleshooting.
}
\description{
A set of input validation functions for internal use that ensure
the passed arguments meet the expected criteria (e.g., type, length, and
value ranges). These functions catch common input errors, provide
meaningful error messages and guidance for correction.

\code{check_df()} ensures that \code{x} is a data frame. If it
is not, an error is thrown with a message indicating the issue.

\code{check_eta()} validates that the input is a numeric value
and, if it is greater than 1, provides a warning about potential
overfitting.

\code{check_length()} ensures that \code{x} contains at least
one element.

\code{check_list()} validates that the input is a list.

\code{check_numeric()} validates that the input is a numeric
value.

\code{check_train()} ensures that the training and test
set names are not the same.

\code{check_prop()} validates that the input is a proportion (a
numeric value between 0 and 1).
}
\details{
Ensures that the input is not empty. If the input has a length of 0,
an error message is raised with a suggestion to check the input.

Ensures that the input is a valid list. If not, an error message is raised.

Ensures that the input is numeric. If not, an error message is raised with a
suggestion to validate the input.
}
\examples{
# Example usage:
\dontrun{check_df(5)}  # Fails with an error message

# Example usage:
\dontrun{check_eta("a")}  # Fails with an error

# Example usage:
\dontrun{check_length(character(0))}  # Fails with an error

# Example usage:
\dontrun{check_list(5)}  # Fails with an error

# Example usage:
\dontrun{check_numeric("a")}  # Fails with an error

# Example usage:
\dontrun{check_train("train", "train")}  # Fails with an error

# Example usage:
\dontrun{check_prop(0)}  # Issues a warning

}
