% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feedback}
\alias{feedback}
\alias{walking_dots}
\alias{color_message}
\alias{walking_colordots}
\title{Utility Functions for Console Feedback and Styling}
\usage{
walking_dots(n = 3, delay = 0.2)

color_message(text, color_code = 32, newline = FALSE)

walking_colordots(n = 3, delay = 0.2, color_code = 32)
}
\arguments{
\item{n}{An integer specifying the number of iterations for dots or
animations. Defaults to 3.}

\item{delay}{A numeric value specifying the delay (in seconds) between
iterations or dots. Defaults to 0.2.}

\item{text}{A character string representing the message to be displayed. Used
in \code{color_message()}.}

\item{color_code}{An integer specifying the ANSI color code for the text.
Defaults to 32 (green).}

\item{newline}{A logical value indicating whether to append a newline after
the message. Defaults to \code{FALSE}.}
}
\value{
All functions are called for their side effects. They do not return
values, but produce visual effects in the console, such as animations or
colored text.
}
\description{
A set of utility functions to enhance the console feedback
during processes. These functions include animated feedback, colored
messages, and combinations of both for a better user experience in the
terminal.

\code{walking_dots()} creates a sequence of animated dots (e.g.,
"...") followed by a "Done" message. The number of dots and the delay can
be customized.

\code{color_message()} prints a message in a specified color
using ANSI color codes. The text and color can be customized.

\code{walking_colordots()} is a combination of
\code{walking_dots()} and \code{color_message()}. It prints animated dots
in a specified color, with a delay between each, and ends with a colored
"Done" message.
}
\examples{
# Example usage:
\dontrun{walking_dots(n = 5, delay = 0.1)}

# Example usage:
\dontrun{color_message("A green message", color_code = 32)}

# Example usage:
\dontrun{walking_colordots(n = 4, delay = 0.15, color_code = 34)}

}
