% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_split.R
\name{random_split}
\alias{random_split}
\title{Randomly Split a Dataset into Two Subsets}
\usage{
random_split(set, prop = 0.7)
}
\arguments{
\item{set}{A data frame or a matrix to be split into subsets.}

\item{prop}{A numeric value between 0 and 1 indicating the proportion of rows
to include in the first subset. Defaults to 0.7.}
}
\value{
A list with two elements:
\describe{
\item{\code{set1}}{The first subset of the dataset, containing
\code{prop} proportion of the rows.}
\item{\code{set2}}{The second subset of the dataset, containing the
remaining rows.}
}
}
\description{
The function takes a dataset and splits it into two subsets randomly, based
on a specified proportion.
}
\examples{
# Load a dataset
data(altmejd)

# A 70-30 simple random split
split <- random_split(altmejd, prop = 0.7)

# Access the subsets
set1 <- split$set1
set2 <- split$set2

}
