% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_split.R
\name{stratified_split}
\alias{stratified_split}
\title{Stratified Split of a Dataset into Two Subsets}
\usage{
stratified_split(set, group, id, prop = 0.7)
}
\arguments{
\item{set}{A data frame to be split into subsets.}

\item{group}{A character string specifying the name of the outcome variable
in \code{set}.}

\item{id}{A character string specifying the name of the (categorical) stratum
variable in \code{set}.}

\item{prop}{A numeric value between 0 and 1 indicating the proportion of rows
to include in the first subset. Defaults to 0.7.}
}
\value{
A list with two elements:
\describe{
\item{\code{set1}}{The first subset of the dataset, containing
approximately \code{prop} proportion of the rows, stratified by the grouping variable.}
\item{\code{set2}}{The second subset of the dataset, containing the
remaining rows.}
}
}
\description{
The function takes a dataset and splits it into two subsets while preserving
the distribution of a specified grouping variable.
}
\examples{
# Load a dataset
data(altmejd)

# Perform a 70-30 stratified split by
split <- stratified_split(altmejd, group = "replicate", id = "eid", prop = 0.7)

# Access the subsets
set1 <- split$set1
set2 <- split$set2

}
