% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAda.R
\name{testAda}
\alias{testAda}
\title{Generate Predictions From a Trained AdaBoost Classifier}
\usage{
testAda(H, data, input_checks = TRUE, verbose = TRUE)
}
\arguments{
\item{H}{A list representing the trained AdaBoost model. Each element of the
list contains a weak learner (e.g., a decision stump) and its corresponding
weight.}

\item{data}{A data frame containing the test dataset for predictions. The
structure should match the training dataset used to generate \code{H}.}

\item{input_checks}{A logical value indicating whether to perform input
validation checks. Defaults to \code{TRUE}.}

\item{verbose}{A logical value specifying whether to display progress
messages and animations. Defaults to \code{TRUE}.}
}
\value{
A numeric vector containing the final predictions from the AdaBoost
model. The predictions are in the form of \code{-1} or \code{1},
corresponding to binary classification outcomes.
}
\description{
This function performs the testing phase of the AdaBoost
algorithm. It processes the trained model, validates input data, extracts
weak learners, and makes predictions or retrodictions on the test dataset.
}
\details{
The function proceeds as follows:
\enumerate{
\item If \code{input_checks} is \code{TRUE}, basic input validation is
performed to ensure that \code{H} is a valid AdaBoost model and \code{data}
is appropriate for predictions.
\item Progress messages and animations are shown if \code{verbose} is
\code{TRUE}.
\item Weak learners (\code{h}) and their weights (\code{a}) are extracted
from \code{H}.
\item Predictions or retrodictions are generated for each weak learner
using the test set.
\item Individual predictions are combined using the weights from the
AdaBoost model to produce the final predictions.
}
}
\examples{
# Example usage:
# Assume `H` is a trained model and `test` is a data frame.
# y_pred <- testAda(H, test_data)

}
