% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainAda.R
\name{trainAda}
\alias{trainAda}
\title{Train a Set of Classification Trees Using AdaBoost}
\usage{
trainAda(formula, data, T, eta, treehypar, input_checks = TRUE, verbose = TRUE)
}
\arguments{
\item{formula}{A \code{\link[stats]{formula}} specifying the relationship
between the outcome and the predictors: \code{outcome ~ predictors}. The
predictors should be included as additive terms (e.g., \code{X1 + X2 + ...}),
Interactions are not supported (see the \code{formula} argument in
\code{\link[rpart]{rpart}}).}

\item{data}{A data frame containing the variables in the model.}

\item{T}{An integer specifying the number of trees.}

\item{eta}{A numeric value representing the learning rate of the algorithm.}

\item{treehypar}{A list of control parameters for decision trees
(passed to \code{\link[rpart]{rpart.control}}).}

\item{input_checks}{A logical value indicating whether to perform input
validation checks. Defaults to \code{TRUE}.}

\item{verbose}{A logical value indicating whether to display verbose output
during the training process. Defaults to \code{TRUE}.}
}
\value{
A list containing the trained weak learners (\code{h}) and their
associated weights (\code{a}). Additional attributes may be included for model
tracking purposes.
}
\description{
The function trains a set of classification trees using
AdaBoost. It is built on top of the \code{rpart} package, so the full range
of tree hyperparameters can be used to fine-tune the trees (see, e.g.,
\code{\link[rpart]{rpart.control}}). In addition, the implementation
allows customization of the number of trees and the learning rate. It also
provides verbose output to track the training progress.
}
\details{
The function implements the AdaBoost algorithm with a specified
number of iterations (\code{T}). It initializes observation weights, trains a
sequence of decision trees, and updates the weights at each iteration based
on prediction errors. A final ensemble of weak learners is produced.

Key steps in the algorithm:
\enumerate{
\item Initialize observation weights.
\item Train a decision tree using the current weights.
\item Compute the weighted classification error and update the observation
weights.
\item Store the weak learner and its associated weight.
}
}
\examples{
\dontrun{
# Example usage:
library(rpart)
data(iris)
# Prepare binary classification data
irisbin <- iris[iris$Species != "setosa", ]
irisbin$Species <- factor(irisbin$Species)

# Set tree hyperparameters
treehypar <- rpart::rpart.control(maxdepth = 1, cp = 0)

# Train AdaBoost model
model <- trainAda(Species ~ ., data = irisbin, T = 10, eta = 1,
treehypar = treehypar)
}

}
